applied styles to pages and components

+++ src/pages/Contact.tsx
@@ -12,6 +12,7 @@ import {
 } from "../shared/apicall"
 import { AppContext, IAppContext } from "../AppContext"
 import EditContactForm from "../components/EditContact"
+import Swal from "sweetalert2"
 
 const ContactPage: React.FC = () => {
     const { contactId } = useParams()
@@ -19,7 +20,7 @@ const ContactPage: React.FC = () => {
         {}
     )
 
-    const navigate = useNavigate()
+    // const navigate = useNavigate()
 
     const { user, setUser } = useContext(
         AppContext
@@ -58,12 +59,34 @@ const ContactPage: React.FC = () => {
     }
     const handleDelete = async (contactId: string) => {
         try {
-            await deleteContact(contactId)
-        } catch (error) {}
+            let cont = await Swal.fire({
+                text: "Are you sure you want to delete this contact?",
+                confirmButtonText: "Continue",
+                showCancelButton: true,
+                icon: "warning",
+                color: "#fff",
+                background: "#59afade9",
+                confirmButtonColor: "#2c3e50",
+                focusConfirm: false,
+            })
+            if (cont.isConfirmed) {
+                await deleteContact(contactId)
+                navigate(`/dashboard`)
+            }
+        } catch (error) {
+            console.log()
+        }
+    }
+
+    const navigate = useNavigate()
+
+    const handleClick = () => {
+        navigate(`/dashboard`)
     }
 
     return (
-        <div className="auth-container">
+        <>
+            {" "}
             <>
                 {editingContact && (
                     <EditContactForm
@@ -72,31 +95,72 @@ const ContactPage: React.FC = () => {
                     />
                 )}
             </>
-            <>
-                <p className="grid-item">
-                    {contact.firstName}
-                </p>
-                <p className="grid-item">
-                    {contact.lastName}
-                </p>
-                <p className="grid-item">{contact.phone}</p>
-                <p className="grid-item">{contact.email}</p>
-                <button
-                    className="contact-button"
-                    onClick={() => openEdit(contact)}
-                >
-                    Edit
-                </button>
-                <button
-                    className="contact-button"
-                    onClick={() =>
-                        handleDelete(contact.contactId!)
-                    }
-                >
-                    Delete
-                </button>
-            </>
-        </div>
+            <button
+                className="contact-button back"
+                onClick={handleClick}
+            >
+                Back
+            </button>
+            <div className="auth-container">
+                <>
+                    <h1>
+                        {contact.firstName &&
+                            contact.firstName[0]}
+                        {contact.lastName &&
+                            contact.lastName[0]}
+                    </h1>
+                    <div className="grid12">
+                        <p>First name : </p>
+                        <p className="grid-item">
+                            {contact.firstName}
+                        </p>
+                    </div>
+
+                    <div className="grid12">
+                        <p>Last name : </p>
+                        <p className="grid-item">
+                            {contact.lastName}
+                        </p>
+                    </div>
+
+                    <div className="grid12">
+                        <p>Phone : </p>
+                        <p className="grid-item">
+                            {contact.phone}
+                        </p>
+                    </div>
+
+                    {contact.email && (
+                        <div className="grid12">
+                            <p>Email : </p>
+                            <p className="grid-item">
+                                {contact.email}
+                            </p>
+                        </div>
+                    )}
+                    <div className="button-div">
+                        <button
+                            className="contact-button"
+                            onClick={() =>
+                                openEdit(contact)
+                            }
+                        >
+                            Edit
+                        </button>
+                        <button
+                            className="contact-button"
+                            onClick={() =>
+                                handleDelete(
+                                    contact.contactId!
+                                )
+                            }
+                        >
+                            Delete
+                        </button>
+                    </div>
+                </>
+            </div>
+        </>
     )
 }
